+++ modules/balancer/controlplane/mod.go
@@ -38,7 +38,7 @@ func NewBalancerModule(
 		zap.Stringer("size", cfg.MemoryRequirements),
 	)
 
-	agent, err := shm.AgentAttach(agentName, cfg.InstanceID, uint(cfg.MemoryRequirements))
+	agent, err := shm.AgentAttach(agentName, cfg.InstanceID, cfg.MemoryRequirements)
 	if err != nil {
 		return nil, fmt.Errorf("failed to attach agent to shared memory: %w", err)
 	}
