+++ modules/decap/controlplane/mod.go
@@ -33,7 +33,7 @@ func NewDecapModule(cfg *Config, log *zap.SugaredLogger) (*DecapModule, error) {
 		zap.Stringer("size", cfg.MemoryRequirements),
 	)
 
-	agent, err := shm.AgentAttach("decap", cfg.InstanceID, uint(cfg.MemoryRequirements))
+	agent, err := shm.AgentAttach("decap", cfg.InstanceID, cfg.MemoryRequirements)
 	if err != nil {
 		return nil, fmt.Errorf("failed to attach agent to shared memory: %w", err)
 	}
