+++ modules/nat64/controlplane/mod.go
@@ -33,7 +33,7 @@ func NewNAT64Module(cfg *Config, log *zap.SugaredLogger) (*NAT64Module, error) {
 		zap.Stringer("size", cfg.MemoryRequirements),
 	)
 
-	agent, err := shm.AgentAttach("nat64", cfg.InstanceID, uint(cfg.MemoryRequirements))
+	agent, err := shm.AgentAttach("nat64", cfg.InstanceID, cfg.MemoryRequirements)
 	if err != nil {
 		return nil, fmt.Errorf("failed to attach agent to shared memory: %w", err)
 	}
