+++ modules/pdump/controlplane/mod.go
@@ -41,7 +41,7 @@ func NewPdumpModule(cfg *Config, log *zap.SugaredLogger) (*PdumpModule, error) {
 		zap.Stringer("size", cfg.MemoryRequirements),
 	)
 
-	agent, err := shm.AgentAttach("pdump", cfg.InstanceID, uint(cfg.MemoryRequirements))
+	agent, err := shm.AgentAttach("pdump", cfg.InstanceID, cfg.MemoryRequirements)
 	if err != nil {
 		return nil, fmt.Errorf("failed to attach agent to shared memory: %w", err)
 	}
